# apps/profile-api/openapi-definition.yml
openapi: 3.0.3
info:
  title: Life Events Profile API
  description: API for Life Events Profile Service
  version: 0.0.1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/profiles/:
    get:
      operationId: indexProfiles
      tags:
        - Profiles
      parameters:
        - schema:
            type: string
          in: query
          name: search
          required: false
          description: If set, the endpoint searches for users whom contain this value in
            either the public name or the email address
        - schema:
            pattern: ^[0-9][0-9]*|undefined$
            default: "0"
            type: string
          in: query
          name: offset
          required: false
          description: Indicates where to start fetching data or how many records to skip,
            defining the initial position within the list
        - schema:
            default: "20"
            pattern: ^([1-9]|100)|undefined$
            type: string
          in: query
          name: limit
          required: false
          description: Indicates the maximum number (100) of items that will be returned
            in a single request
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        publicName:
                          type: string
                        email:
                          format: email
                          type: string
                        primaryUserId:
                          type: string
                        safeLevel:
                          type: number
                        preferredLanguage:
                          type: string
                          enum:
                            - en
                            - ga
                          default: en
                        createdAt:
                          format: date-time
                          type: string
                        updatedAt:
                          format: date-time
                          type: string
                      required:
                        - id
                        - publicName
                        - email
                        - primaryUserId
                  metadata:
                    type: object
                    properties:
                      links:
                        description: Object containing the links to the related endpoints
                        type: object
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the request itself
                                type: string
                          next:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the next page of results in a paginated response.
                                  If there are no more results, this field may
                                  be omitted
                                type: string
                          prev:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the previous page of results in a paginated
                                  response. If there are no more results, this
                                  field may be omitted
                                type: string
                          first:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the first page of results in a paginated response
                                type: string
                          last:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the first page of results in a paginated response
                                type: string
                          pages:
                            description: It may contain a list of other useful URLs, e.g. one entry for
                              page:'page 1', 'page 2'
                            type: object
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        required:
                          - self
                          - first
                          - last
                          - pages
                      totalCount:
                        description: Number representing the total number of available items
                        type: number
                required:
                  - data
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    description: Code used to categorize the error
                    type: string
                  detail:
                    description: Description of the error
                    type: string
                  requestId:
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                    type: string
                  name:
                    description: Name of the error type
                    type: string
                  validation:
                    description: List of the validation errors
                    type: array
                    items:
                      type: object
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                      required:
                        - fieldName
                        - message
                  validationContext:
                    type: string
                  statusCode:
                    type: number
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    description: Code used to categorize the error
                    type: string
                  detail:
                    description: Description of the error
                    type: string
                  requestId:
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                    type: string
                  name:
                    description: Name of the error type
                    type: string
                  validation:
                    description: List of the validation errors
                    type: array
                    items:
                      type: object
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                      required:
                        - fieldName
                        - message
                  validationContext:
                    type: string
                  statusCode:
                    type: number
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
  /api/v1/profiles/import-profiles:
    post:
      operationId: importProfiles
      tags:
        - Profiles
      requestBody:
        content:
          application/json: &a1
            schema:
              type: object
              properties:
                profiles:
                  minItems: 1
                  type: array
                  items:
                    type: object
                    properties:
                      email:
                        format: email
                        type: string
                      firstName:
                        type: string
                      lastName:
                        type: string
                      city:
                        type: string
                      address:
                        type: string
                      phone:
                        type: string
                      dateOfBirth:
                        format: date
                        type: string
                      ppsn:
                        type: string
                      preferredLanguage:
                        type: string
                        enum:
                          - en
                          - ga
                        default: en
                    required:
                      - email
                      - firstName
                      - lastName
                file: {}
          multipart/form-data: *a1
          text/csv: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  profileImportId:
                    type: string
                required:
                  - status
                  - profileImportId
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    description: Code used to categorize the error
                    type: string
                  detail:
                    description: Description of the error
                    type: string
                  requestId:
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                    type: string
                  name:
                    description: Name of the error type
                    type: string
                  validation:
                    description: List of the validation errors
                    type: array
                    items:
                      type: object
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                      required:
                        - fieldName
                        - message
                  validationContext:
                    type: string
                  statusCode:
                    type: number
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    description: Code used to categorize the error
                    type: string
                  detail:
                    description: Description of the error
                    type: string
                  requestId:
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                    type: string
                  name:
                    description: Name of the error type
                    type: string
                  validation:
                    description: List of the validation errors
                    type: array
                    items:
                      type: object
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                      required:
                        - fieldName
                        - message
                  validationContext:
                    type: string
                  statusCode:
                    type: number
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
  /api/v1/profiles/select-profiles:
    get:
      operationId: selectProfiles
      tags:
        - Profiles
      parameters:
        - schema:
            pattern: ^[a-zA-Z0-9-]+(,[a-zA-Z0-9-]+)*$
            type: string
          in: query
          name: ids
          required: true
          description: Comma-separated list of profile IDs
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        publicName:
                          type: string
                        email:
                          format: email
                          type: string
                        primaryUserId:
                          type: string
                        safeLevel:
                          type: number
                        preferredLanguage:
                          type: string
                          enum:
                            - en
                            - ga
                          default: en
                        createdAt:
                          format: date-time
                          type: string
                        updatedAt:
                          format: date-time
                          type: string
                        details:
                          type: object
                          properties:
                            email:
                              format: email
                              type: string
                            firstName:
                              type: string
                            lastName:
                              type: string
                            city:
                              type: string
                            address:
                              type: string
                            phone:
                              type: string
                            dateOfBirth:
                              format: date
                              type: string
                            ppsn:
                              type: string
                            preferredLanguage:
                              type: string
                              enum:
                                - en
                                - ga
                              default: en
                          required:
                            - email
                            - firstName
                            - lastName
                      required:
                        - id
                        - publicName
                        - email
                        - primaryUserId
                  metadata:
                    type: object
                    properties:
                      links:
                        description: Object containing the links to the related endpoints
                        type: object
                        properties:
                          self:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the request itself
                                type: string
                          next:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the next page of results in a paginated response.
                                  If there are no more results, this field may
                                  be omitted
                                type: string
                          prev:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the previous page of results in a paginated
                                  response. If there are no more results, this
                                  field may be omitted
                                type: string
                          first:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the first page of results in a paginated response
                                type: string
                          last:
                            type: object
                            properties:
                              href:
                                description: URL pointing to the first page of results in a paginated response
                                type: string
                          pages:
                            description: It may contain a list of other useful URLs, e.g. one entry for
                              page:'page 1', 'page 2'
                            type: object
                            additionalProperties:
                              type: object
                              properties:
                                href:
                                  type: string
                        required:
                          - self
                          - first
                          - last
                          - pages
                      totalCount:
                        description: Number representing the total number of available items
                        type: number
                required:
                  - data
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    description: Code used to categorize the error
                    type: string
                  detail:
                    description: Description of the error
                    type: string
                  requestId:
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                    type: string
                  name:
                    description: Name of the error type
                    type: string
                  validation:
                    description: List of the validation errors
                    type: array
                    items:
                      type: object
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                      required:
                        - fieldName
                        - message
                  validationContext:
                    type: string
                  statusCode:
                    type: number
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    description: Code used to categorize the error
                    type: string
                  detail:
                    description: Description of the error
                    type: string
                  requestId:
                    description: Unique request id. This one will be used to troubleshoot the
                      problems
                    type: string
                  name:
                    description: Name of the error type
                    type: string
                  validation:
                    description: List of the validation errors
                    type: array
                    items:
                      type: object
                      properties:
                        fieldName:
                          type: string
                        message:
                          type: string
                      required:
                        - fieldName
                        - message
                  validationContext:
                    type: string
                  statusCode:
                    type: number
                required:
                  - code
                  - detail
                  - requestId
                  - name
                  - statusCode
...
security:
  - bearerAuth: []